chore: rename files for clarity and delete some #271
