+++ apps/mai-sc/src/ui/button-scroll-to-bottom.tsx
@@ -21,7 +21,7 @@ export function ButtonScrollToBottom({
       variant="outline"
       size="icon"
       className={cn(
-        'absolute right-4 top-1 z-10 bg-background transition-opacity duration-300 sm:right-8 md:top-2',
+        'absolute right-4 top-1 z-50 bg-background transition-opacity duration-300 sm:right-8 md:top-2',
         isAtBottom ? 'opacity-0' : 'opacity-100',
         className,
       )}
