+++ apps/mai-sc/src/ui/logo.tsx
