+++ apps/mai-sc/src/ui/mai/mai-chat.tsx
@@ -1,13 +1,13 @@
 'use client';
 
-import { ChatMessages } from '#/ui/mai/chat-messages';
-import { ChatPanel } from '#/ui/mai/chat-panel';
+import { MaiChatMessages } from '#/ui/mai/mai-chat-messages';
+import { MaiChatPanel } from '#/ui/mai/mai-chat-panel';
 
-export function Chat() {
+export function MaiChat() {
   return (
     <div className="no-scrollbar mx-auto flex w-full max-w-3xl flex-col space-y-3 overflow-auto px-8 pb-14 pt-6 md:space-y-4 md:px-12 md:pb-24 md:pt-8">
-      <ChatMessages />
-      <ChatPanel />
+      <MaiChatMessages />
+      <MaiChatPanel />
     </div>
   );
 }
